added gradio multi-step process w/o initial frame generation
