+++ .gitignore
@@ -129,4 +129,5 @@ dmypy.json
 .pyre/
 
 ###
+.vscode/
 *.pth
