+++ app_text_to_video.py
@@ -1,5 +1,5 @@
 import gradio as gr
-from pipelines.model import ControlAnimationModel
+from models.model import ControlAnimationModel
 import os
 from utils.hf_utils import get_model_list
