+++ models/unet_2d_condition_flax.py
@@ -11,6 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+# CUSTOM VERSION OF https://github.com/huggingface/diffusers/blob/main/src/diffusers/models/unet_2d_condition_flax.py
+
 from typing import Tuple, Union
 
 import flax
@@ -23,84 +26,31 @@ from diffusers.configuration_utils import ConfigMixin, flax_register_to_config
 from diffusers.utils import BaseOutput
 from diffusers.models.embeddings_flax import FlaxTimestepEmbedding, FlaxTimesteps
 from diffusers.models.modeling_flax_utils import FlaxModelMixin
-from models.flax_unet_2d_condition import (
+from diffusers.models.unet_2d_blocks_flax import (
     FlaxCrossAttnDownBlock2D,
+    FlaxCrossAttnUpBlock2D,
     FlaxDownBlock2D,
     FlaxUNetMidBlock2DCrossAttn,
+    FlaxUpBlock2D,
 )
 
 
 @flax.struct.dataclass
-class FlaxControlNetOutput(BaseOutput):
-    down_block_res_samples: jnp.ndarray
-    mid_block_res_sample: jnp.ndarray
+class FlaxUNet2DConditionOutput(BaseOutput):
+    """
+    Args:
+        sample (`jnp.ndarray` of shape `(batch_size, num_channels, height, width)`):
+            Hidden states conditioned on `encoder_hidden_states` input. Output of last layer of model.
+    """
 
-
-class FlaxControlNetConditioningEmbedding(nn.Module):
-    conditioning_embedding_channels: int
-    block_out_channels: Tuple[int] = (16, 32, 96, 256)
-    dtype: jnp.dtype = jnp.float32
-
-    def setup(self):
-        self.conv_in = nn.Conv(
-            self.block_out_channels[0],
-            kernel_size=(3, 3),
-            padding=((1, 1), (1, 1)),
-            dtype=self.dtype,
-        )
-
-        blocks = []
-        for i in range(len(self.block_out_channels) - 1):
-            channel_in = self.block_out_channels[i]
-            channel_out = self.block_out_channels[i + 1]
-            conv1 = nn.Conv(
-                channel_in,
-                kernel_size=(3, 3),
-                padding=((1, 1), (1, 1)),
-                dtype=self.dtype,
-            )
-            blocks.append(conv1)
-            conv2 = nn.Conv(
-                channel_out,
-                kernel_size=(3, 3),
-                strides=(2, 2),
-                padding=((1, 1), (1, 1)),
-                dtype=self.dtype,
-            )
-            blocks.append(conv2)
-        self.blocks = blocks
-
-        self.conv_out = nn.Conv(
-            self.conditioning_embedding_channels,
-            kernel_size=(3, 3),
-            padding=((1, 1), (1, 1)),
-            kernel_init=nn.initializers.zeros_init(),
-            bias_init=nn.initializers.zeros_init(),
-            dtype=self.dtype,
-        )
-
-    def __call__(self, conditioning):
-        embedding = self.conv_in(conditioning)
-        embedding = nn.silu(embedding)
-
-        for block in self.blocks:
-            embedding = block(embedding)
-            embedding = nn.silu(embedding)
-
-        embedding = self.conv_out(embedding)
-
-        return embedding
+    sample: jnp.ndarray
 
 
 @flax_register_to_config
-class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
+class FlaxUNet2DConditionModel(nn.Module, FlaxModelMixin, ConfigMixin):
     r"""
-    Quoting from https://arxiv.org/abs/2302.05543: "Stable Diffusion uses a pre-processing method similar to VQ-GAN
-    [11] to convert the entire dataset of 512 × 512 images into smaller 64 × 64 “latent images” for stabilized
-    training. This requires ControlNets to convert image-based conditions to 64 × 64 feature space to match the
-    convolution size. We use a tiny network E(·) of four convolution layers with 4 × 4 kernels and 2 × 2 strides
-    (activated by ReLU, channels are 16, 32, 64, 128, initialized with Gaussian weights, trained jointly with the full
-    model) to encode image-space conditions ... into feature maps ..."
+    FlaxUNet2DConditionModel is a conditional 2D UNet model that takes in a noisy sample, conditional state, and a
+    timestep and returns sample shaped output.
 
     This model inherits from [`FlaxModelMixin`]. Check the superclass documentation for the generic methods the library
     implements for all the models (such as downloading or saving, etc.)
@@ -120,9 +70,14 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
             The size of the input sample.
         in_channels (`int`, *optional*, defaults to 4):
             The number of channels in the input sample.
+        out_channels (`int`, *optional*, defaults to 4):
+            The number of channels in the output.
         down_block_types (`Tuple[str]`, *optional*, defaults to `("CrossAttnDownBlock2D", "CrossAttnDownBlock2D", "CrossAttnDownBlock2D", "DownBlock2D")`):
             The tuple of downsample blocks to use. The corresponding class names will be: "FlaxCrossAttnDownBlock2D",
             "FlaxCrossAttnDownBlock2D", "FlaxCrossAttnDownBlock2D", "FlaxDownBlock2D"
+        up_block_types (`Tuple[str]`, *optional*, defaults to `("UpBlock2D", "CrossAttnUpBlock2D", "CrossAttnUpBlock2D", "CrossAttnUpBlock2D",)`):
+            The tuple of upsample blocks to use. The corresponding class names will be: "FlaxUpBlock2D",
+            "FlaxCrossAttnUpBlock2D", "FlaxCrossAttnUpBlock2D", "FlaxCrossAttnUpBlock2D"
         block_out_channels (`Tuple[int]`, *optional*, defaults to `(320, 640, 1280, 1280)`):
             The tuple of output channels for each block.
         layers_per_block (`int`, *optional*, defaults to 2):
@@ -136,20 +91,25 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
         flip_sin_to_cos (`bool`, *optional*, defaults to `True`):
             Whether to flip the sin to cos in the time embedding.
         freq_shift (`int`, *optional*, defaults to 0): The frequency shift to apply to the time embedding.
-        controlnet_conditioning_channel_order (`str`, *optional*, defaults to `rgb`):
-            The channel order of conditional image. Will convert it to `rgb` if it's `bgr`
-        conditioning_embedding_out_channels (`tuple`, *optional*, defaults to `(16, 32, 96, 256)`):
-            The tuple of output channel for each block in conditioning_embedding layer
-
+        use_memory_efficient_attention (`bool`, *optional*, defaults to `False`):
+            enable memory efficient attention https://arxiv.org/abs/2112.05682
 
     """
+
     sample_size: int = 32
     in_channels: int = 4
+    out_channels: int = 4
     down_block_types: Tuple[str] = (
         "CrossAttnDownBlock2D",
         "CrossAttnDownBlock2D",
         "CrossAttnDownBlock2D",
         "DownBlock2D",
+    )
+    up_block_types: Tuple[str] = (
+        "UpBlock2D",
+        "CrossAttnUpBlock2D",
+        "CrossAttnUpBlock2D",
+        "CrossAttnUpBlock2D",
     )
     only_cross_attention: Union[bool, Tuple[bool]] = False
     block_out_channels: Tuple[int] = (320, 640, 1280, 1280)
@@ -161,22 +121,21 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
     dtype: jnp.dtype = jnp.float32
     flip_sin_to_cos: bool = True
     freq_shift: int = 0
-    controlnet_conditioning_channel_order: str = "rgb"
-    conditioning_embedding_out_channels: Tuple[int] = (16, 32, 96, 256)
+    use_memory_efficient_attention: bool = False
 
     def init_weights(self, rng: jax.random.KeyArray) -> FrozenDict:
         # init input tensors
         sample_shape = (1, self.in_channels, self.sample_size, self.sample_size)
         sample = jnp.zeros(sample_shape, dtype=jnp.float32)
         timesteps = jnp.ones((1,), dtype=jnp.int32)
-        encoder_hidden_states = jnp.zeros((1, 1, self.cross_attention_dim), dtype=jnp.float32)
-        controlnet_cond_shape = (1, 3, self.sample_size * 8, self.sample_size * 8)
-        controlnet_cond = jnp.zeros(controlnet_cond_shape, dtype=jnp.float32)
+        encoder_hidden_states = jnp.zeros(
+            (1, 1, self.cross_attention_dim), dtype=jnp.float32
+        )
 
         params_rng, dropout_rng = jax.random.split(rng)
         rngs = {"params": params_rng, "dropout": dropout_rng}
 
-        return self.init(rngs, sample, timesteps, encoder_hidden_states, controlnet_cond)["params"]
+        return self.init(rngs, sample, timesteps, encoder_hidden_states)["params"]
 
     def setup(self):
         block_out_channels = self.block_out_channels
@@ -193,14 +152,11 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
 
         # time
         self.time_proj = FlaxTimesteps(
-            block_out_channels[0], flip_sin_to_cos=self.flip_sin_to_cos, freq_shift=self.config.freq_shift
+            block_out_channels[0],
+            flip_sin_to_cos=self.flip_sin_to_cos,
+            freq_shift=self.config.freq_shift,
         )
         self.time_embedding = FlaxTimestepEmbedding(time_embed_dim, dtype=self.dtype)
-
-        self.controlnet_cond_embedding = FlaxControlNetConditioningEmbedding(
-            conditioning_embedding_channels=block_out_channels[0],
-            block_out_channels=self.conditioning_embedding_out_channels,
-        )
 
         only_cross_attention = self.only_cross_attention
         if isinstance(only_cross_attention, bool):
@@ -212,20 +168,7 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
 
         # down
         down_blocks = []
-        controlnet_down_blocks = []
-
         output_channel = block_out_channels[0]
-
-        controlnet_block = nn.Conv(
-            output_channel,
-            kernel_size=(1, 1),
-            padding="VALID",
-            kernel_init=nn.initializers.zeros_init(),
-            bias_init=nn.initializers.zeros_init(),
-            dtype=self.dtype,
-        )
-        controlnet_down_blocks.append(controlnet_block)
-
         for i, down_block_type in enumerate(self.down_block_types):
             input_channel = output_channel
             output_channel = block_out_channels[i]
@@ -241,6 +184,7 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
                     add_downsample=not is_final_block,
                     use_linear_projection=self.use_linear_projection,
                     only_cross_attention=only_cross_attention[i],
+                    use_memory_efficient_attention=self.use_memory_efficient_attention,
                     dtype=self.dtype,
                 )
             else:
@@ -254,48 +198,69 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
                 )
 
             down_blocks.append(down_block)
-
-            for _ in range(self.layers_per_block):
-                controlnet_block = nn.Conv(
-                    output_channel,
-                    kernel_size=(1, 1),
-                    padding="VALID",
-                    kernel_init=nn.initializers.zeros_init(),
-                    bias_init=nn.initializers.zeros_init(),
-                    dtype=self.dtype,
-                )
-                controlnet_down_blocks.append(controlnet_block)
-
-            if not is_final_block:
-                controlnet_block = nn.Conv(
-                    output_channel,
-                    kernel_size=(1, 1),
-                    padding="VALID",
-                    kernel_init=nn.initializers.zeros_init(),
-                    bias_init=nn.initializers.zeros_init(),
-                    dtype=self.dtype,
-                )
-                controlnet_down_blocks.append(controlnet_block)
-
         self.down_blocks = down_blocks
-        self.controlnet_down_blocks = controlnet_down_blocks
 
         # mid
-        mid_block_channel = block_out_channels[-1]
         self.mid_block = FlaxUNetMidBlock2DCrossAttn(
-            in_channels=mid_block_channel,
+            in_channels=block_out_channels[-1],
             dropout=self.dropout,
             attn_num_head_channels=attention_head_dim[-1],
             use_linear_projection=self.use_linear_projection,
+            use_memory_efficient_attention=self.use_memory_efficient_attention,
             dtype=self.dtype,
         )
 
-        self.controlnet_mid_block = nn.Conv(
-            mid_block_channel,
-            kernel_size=(1, 1),
-            padding="VALID",
-            kernel_init=nn.initializers.zeros_init(),
-            bias_init=nn.initializers.zeros_init(),
+        # up
+        up_blocks = []
+        reversed_block_out_channels = list(reversed(block_out_channels))
+        reversed_attention_head_dim = list(reversed(attention_head_dim))
+        only_cross_attention = list(reversed(only_cross_attention))
+        output_channel = reversed_block_out_channels[0]
+        for i, up_block_type in enumerate(self.up_block_types):
+            prev_output_channel = output_channel
+            output_channel = reversed_block_out_channels[i]
+            input_channel = reversed_block_out_channels[
+                min(i + 1, len(block_out_channels) - 1)
+            ]
+
+            is_final_block = i == len(block_out_channels) - 1
+
+            if up_block_type == "CrossAttnUpBlock2D":
+                up_block = FlaxCrossAttnUpBlock2D(
+                    in_channels=input_channel,
+                    out_channels=output_channel,
+                    prev_output_channel=prev_output_channel,
+                    num_layers=self.layers_per_block + 1,
+                    attn_num_head_channels=reversed_attention_head_dim[i],
+                    add_upsample=not is_final_block,
+                    dropout=self.dropout,
+                    use_linear_projection=self.use_linear_projection,
+                    only_cross_attention=only_cross_attention[i],
+                    use_memory_efficient_attention=self.use_memory_efficient_attention,
+                    dtype=self.dtype,
+                )
+            else:
+                up_block = FlaxUpBlock2D(
+                    in_channels=input_channel,
+                    out_channels=output_channel,
+                    prev_output_channel=prev_output_channel,
+                    num_layers=self.layers_per_block + 1,
+                    add_upsample=not is_final_block,
+                    dropout=self.dropout,
+                    dtype=self.dtype,
+                )
+
+            up_blocks.append(up_block)
+            prev_output_channel = output_channel
+        self.up_blocks = up_blocks
+
+        # out
+        self.conv_norm_out = nn.GroupNorm(num_groups=32, epsilon=1e-5)
+        self.conv_out = nn.Conv(
+            self.out_channels,
+            kernel_size=(3, 3),
+            strides=(1, 1),
+            padding=((1, 1), (1, 1)),
             dtype=self.dtype,
         )
 
@@ -304,18 +269,16 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
         sample,
         timesteps,
         encoder_hidden_states,
-        controlnet_cond,
-        conditioning_scale: float = 1.0,
+        down_block_additional_residuals=None,
+        mid_block_additional_residual=None,
         return_dict: bool = True,
         train: bool = False,
-    ) -> Union[FlaxControlNetOutput, Tuple]:
+    ) -> Union[FlaxUNet2DConditionOutput, Tuple]:
         r"""
         Args:
             sample (`jnp.ndarray`): (batch, channel, height, width) noisy inputs tensor
             timestep (`jnp.ndarray` or `float` or `int`): timesteps
             encoder_hidden_states (`jnp.ndarray`): (batch_size, sequence_length, hidden_size) encoder hidden states
-            controlnet_cond (`jnp.ndarray`): (batch, channel, height, width) the conditional input tensor
-            conditioning_scale: (`float`) the scale factor for controlnet outputs
             return_dict (`bool`, *optional*, defaults to `True`):
                 Whether or not to return a [`models.unet_2d_condition_flax.FlaxUNet2DConditionOutput`] instead of a
                 plain tuple.
@@ -327,10 +290,6 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
             [`~models.unet_2d_condition_flax.FlaxUNet2DConditionOutput`] if `return_dict` is True, otherwise a `tuple`.
             When returning a tuple, the first element is the sample tensor.
         """
-        channel_order = self.controlnet_conditioning_channel_order
-        if channel_order == "bgr":
-            controlnet_cond = jnp.flip(controlnet_cond, axis=1)
-
         # 1. time
         if not isinstance(timesteps, jnp.ndarray):
             timesteps = jnp.array([timesteps], dtype=jnp.int32)
@@ -345,39 +304,65 @@ class FlaxControlNetModel(nn.Module, FlaxModelMixin, ConfigMixin):
         sample = jnp.transpose(sample, (0, 2, 3, 1))
         sample = self.conv_in(sample)
 
-        controlnet_cond = jnp.transpose(controlnet_cond, (0, 2, 3, 1))
-        controlnet_cond = self.controlnet_cond_embedding(controlnet_cond)
-        sample += controlnet_cond
-
         # 3. down
         down_block_res_samples = (sample,)
         for down_block in self.down_blocks:
             if isinstance(down_block, FlaxCrossAttnDownBlock2D):
-                sample, res_samples = down_block(sample, t_emb, encoder_hidden_states, deterministic=not train)
+                sample, res_samples = down_block(
+                    sample, t_emb, encoder_hidden_states, deterministic=not train
+                )
             else:
                 sample, res_samples = down_block(sample, t_emb, deterministic=not train)
             down_block_res_samples += res_samples
 
+        if down_block_additional_residuals is not None:
+            new_down_block_res_samples = ()
+
+            for down_block_res_sample, down_block_additional_residual in zip(
+                down_block_res_samples, down_block_additional_residuals
+            ):
+                down_block_res_sample += down_block_additional_residual
+                new_down_block_res_samples += (down_block_res_sample,)
+
+            down_block_res_samples = new_down_block_res_samples
+
         # 4. mid
-        sample = self.mid_block(sample, t_emb, encoder_hidden_states, deterministic=not train)
+        sample = self.mid_block(
+            sample, t_emb, encoder_hidden_states, deterministic=not train
+        )
 
-        # 5. contronet blocks
-        controlnet_down_block_res_samples = ()
-        for down_block_res_sample, controlnet_block in zip(down_block_res_samples, self.controlnet_down_blocks):
-            down_block_res_sample = controlnet_block(down_block_res_sample)
-            controlnet_down_block_res_samples += (down_block_res_sample,)
+        if mid_block_additional_residual is not None:
+            sample += mid_block_additional_residual
 
-        down_block_res_samples = controlnet_down_block_res_samples
+        # 5. up
+        for up_block in self.up_blocks:
+            res_samples = down_block_res_samples[-(self.layers_per_block + 1) :]
+            down_block_res_samples = down_block_res_samples[
+                : -(self.layers_per_block + 1)
+            ]
+            if isinstance(up_block, FlaxCrossAttnUpBlock2D):
+                sample = up_block(
+                    sample,
+                    temb=t_emb,
+                    encoder_hidden_states=encoder_hidden_states,
+                    res_hidden_states_tuple=res_samples,
+                    deterministic=not train,
+                )
+            else:
+                sample = up_block(
+                    sample,
+                    temb=t_emb,
+                    res_hidden_states_tuple=res_samples,
+                    deterministic=not train,
+                )
 
-        mid_block_res_sample = self.controlnet_mid_block(sample)
-
-        # 6. scaling
-        down_block_res_samples = [sample * conditioning_scale for sample in down_block_res_samples]
-        mid_block_res_sample *= conditioning_scale
+        # 6. post-process
+        sample = self.conv_norm_out(sample)
+        sample = nn.silu(sample)
+        sample = self.conv_out(sample)
+        sample = jnp.transpose(sample, (0, 3, 1, 2))
 
         if not return_dict:
-            return (down_block_res_samples, mid_block_res_sample)
+            return (sample,)
 
-        return FlaxControlNetOutput(
-            down_block_res_samples=down_block_res_samples, mid_block_res_sample=mid_block_res_sample
-        )
+        return FlaxUNet2DConditionOutput(sample=sample)
